requirements1.txt
torch
pyannote.audio
ipython
torchaudio
soundfile
pystoi
transformers
